fix java value incorrect types.

+++ slow-interpreter/src/instructions/ldc.rs
@@ -42,7 +42,7 @@ pub fn create_string_on_stack(jvm: &JVMState, res_string: String) {
         current_loader.clone(),
     );
     let str_as_vec = res_string.chars();
-    let chars: Vec<JavaValue> = str_as_vec.map(|x| { JavaValue::Char(x) }).collect();
+    let chars: Vec<JavaValue> = str_as_vec.map(|x| { JavaValue::Char(x as u16) }).collect();
     push_new_object(jvm, current_frame, &string_class,None);//todo what if stack overflows here?
     let string_object = current_frame.pop();
     let mut args = vec![string_object.clone()];
